Move sundial helper functions to own file
