+++ src/hooks/useSundial/index.ts
@@ -3,10 +3,11 @@ import { SundialMachine } from './machine';
 import { useMemo, useEffect } from 'react';
 import gql from 'graphql-tag';
 import { useApolloClient } from 'react-apollo-hooks';
-import { SunlightWindow } from '../../definitions';
 import useLocalApolloQuery from '../useLocalApolloQuery';
+import { TimeOfDay } from './types';
+import { localDateString, pluckSunlightWindow } from './util';
 
-type SundialState = 'calibrating' | 'beforeSunrise' | 'day' | 'afterSunset';
+type SundialState = 'calibrating' | TimeOfDay;
 
 const SUNLIGHT_WINDOWS_QUERY = gql`
   query SunlightWindows(
@@ -37,45 +38,6 @@ const SUNDIAL_QUERY = gql`
     }
   }
 `;
-
-function dateFromUtcString(utcString: string): Date {
-  const [dateString, timeString] = utcString.split('T');
-  const [fullYear, month, date] = dateString.split('-');
-  const [hour, minute, secondString] = timeString.split(':');
-  const [second, microsecond] = secondString.split('.');
-
-  const utcDate = Date.UTC(
-    parseInt(fullYear),
-    parseInt(month) - 1,
-    parseInt(date),
-    parseInt(hour),
-    parseInt(minute),
-    parseInt(second),
-    microsecond ? parseInt(microsecond) / 1000 : 0
-  );
-  return new Date(utcDate);
-}
-
-function pluckSunlightWindow({
-  sunriseUtc,
-  sunsetUtc
-}: {
-  sunriseUtc: string;
-  sunsetUtc: string;
-}): SunlightWindow {
-  return {
-    sunrise: dateFromUtcString(sunriseUtc),
-    sunset: dateFromUtcString(sunsetUtc)
-  };
-}
-
-function twoDigits(number: number): string {
-  return `0${number}`.slice(-2);
-}
-
-function localDateString(date: Date): string {
-  return `${date.getFullYear()}-${twoDigits(date.getMonth() + 1)}-${twoDigits(date.getDate())}`;
-}
 
 function getVariables() {
   const today = new Date();
