+++ src/hooks/useSundial/types.ts
@@ -1,0 +1,1 @@
+export type TimeOfDay = 'beforeSunrise' | 'day' | 'afterSunset';
